removed crazy console logs and added very early em2 support

+++ EpicMickeyLauncher/src/routes/addgame.svelte
@@ -1,11 +1,11 @@
 <script>
   import { open } from "@tauri-apps/api/dialog";
   import { exists } from "@tauri-apps/api/fs";
-  import { ReadJSON, WriteFile, WriteToJSON } from "./library/configfiles.js";
+  import { ReadJSON, WriteFile, WriteToJSON, ReadFile, FileExists } from "./library/configfiles.js";
   let addgamedumpDiv;
   let dumpFound;
   let error = "";
-
+  let gametype = ""
   let path;
 
   async function AddGameDump() {
@@ -17,6 +17,20 @@
     path = selectedPath;
     let folderExists = await exists(path + "/DATA");
     if (folderExists) {
+
+       let verdat = await FileExists(path + "/DATA/files/VERSIONDATA.TXT")
+
+
+       if(verdat)
+       {
+     //EM1
+      gametype = "EM1"
+       }
+       else{
+     //EM2
+gametype = "EM2"
+       }
+       
       addgamedumpDiv.style.display = "none";
       dumpFound.style.display = "block";
     } else {
@@ -27,13 +41,16 @@
   async function Continue() {
     let jsonData = await ReadJSON("games.json");
 
-    let newData = { path: path + "/DATA", game: "EM1", platform: "wii" }
+    let newData = { path: path + "/DATA", game: gametype, platform: "wii" }
 
     jsonData.push(newData);
 
     await WriteToJSON(JSON.stringify(jsonData), "games.json");
- 
-    await WriteFile("[]", path + "/DATA/EMLMods.json")
+
+    if(FileExists(path + "/DATA/EMLMods.json"))
+    {
+      await WriteFile("[]", path + "/DATA/EMLMods.json")
+    }
 
     window.open("#/", "_self");
   }
